connection funcionando
