+++ modulo5/ToDoList/package.json
@@ -5,23 +5,24 @@
   "main": "index.js",
   "scripts": {
     "dev": "ts-node-dev ./src/index.ts",
-    "start": "tsc && node ./build/index.js",
+    "start": "clear && ts-node-dev ./src/index.ts",
     "post": "ts-node-dev ./src/endpoints/postUser.ts"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@types/knex": "^0.16.1",
     "cors": "^2.8.5",
     "dotenv": "^16.0.1",
     "express": "^4.18.1",
     "knex": "^2.1.0",
-    "mysql": "^2.18.1"
+    "mysql": "^2.18.1",
+    "uuid": "^8.3.2"
   },
   "devDependencies": {
     "@types/cors": "^2.8.12",
     "@types/express": "^4.17.13",
+    "@types/uuid": "^8.3.4",
     "ts-node-dev": "^2.0.0",
     "typescript": "^4.7.4"
   }
