+++ modulo5/ToDoList/src/index.ts
@@ -1,33 +1,23 @@
-import express from "express";
-import knex from "knex";
-import cors from "cors";
-import dotenv from "dotenv";
-import { AddressInfo } from "net";
+import { codes, messages } from './status';
+import connection from "./connection";
+import app from "./app";
+import { Request, Response } from "express";
+import { v4 as generateId } from 'uuid'
 
-dotenv.config();
 
-export const connection = knex({
-	client: "mysql",
-	connection: {
-    host: process.env.DB_HOST,
-    port: 3306,
-    user: process.env.DB_USER,
-    password: process.env.DB_PASS,
-    database: process.env.DB_NAME
-  }
+app.get("/test", async (req: Request, res: Response): Promise<any> => {
+  try {
+  /*  const result = await connection() */
+   console.log();
+   res.status(200).send("projeto funcionando")
+  } catch (error:any) {}
 });
 
-export const app = express();
-app.use(express.json());
-app.use(cors());
+app.post("/user", async (req: Request, res: Response): Promise<any> => {
+   try {
+      
+   } catch (erro:any) {
+      
+   }
 
-
-
-const server = app.listen(process.env.PORT || 3003, () => {
-    if (server) {
-       const address = server.address() as AddressInfo;
-       console.log(`Server is running in PROJETO TODOLIST  http://localhost: ${address.port}`);
-    } else {
-       console.error(`Failure upon starting server.`);
-    }
 });
